add keyboard handling. use forwardrefs. fix close on swipercontext change. add touch to close
